Added authorization in blog data

+++ controllers/blogControllers.js
@@ -45,6 +45,16 @@ const getSingleBlog = async (req, res) => {
   @FUNCTION - Add a new blog
 */
 const createNewBlog = async (req, res) => {
+  // Title is required
+  if (!req.body.title)
+    return res.status(400).json({
+      success: false,
+      message: "Title is required",
+    });
+
+  // Logged in user is the author
+  req.body.author = req.loggedInUser;
+
   // Create and save new blog in DB
   const newBlog = new Blog(req.body);
   await newBlog.save();
@@ -65,7 +75,21 @@ const editBlog = async (req, res) => {
   if (!checkId(req.params.blogId))
     return res.status(404).json({ success: false, message: "Blog not found" });
 
-  // Filter to-be edited detials
+  // Find the blog from DB
+  const foundBlog = await Blog.findById(req.params.blogId);
+
+  // Blog not found
+  if (!foundBlog)
+    return res.status(404).json({ success: false, message: "Blog not found" });
+
+  // Logged In user should be author of blog
+  if (req.loggedInUser != foundBlog.author)
+    return res.status(401).json({
+      success: false,
+      message: "Unauthorized",
+    });
+
+  // Filter to-be edited details
   const editedDetails = {};
   editedDetails.title = req.body.title;
   editedDetails.content = req.body.content;
@@ -76,9 +100,6 @@ const editBlog = async (req, res) => {
     editedDetails,
     { returnDocument: "after" }
   );
-
-  if (!updatedBlog)
-    return res.status(404).json({ success: false, message: "Blog not found" });
 
   return res.status(200).json({
     success: true,
@@ -96,12 +117,22 @@ const deleteBlog = async (req, res) => {
   if (!checkId(req.params.blogId))
     return res.status(404).json({ success: false, message: "Blog not found" });
 
-  // Delete blog from DB
-  const deletedBlog = await Blog.findByIdAndDelete(req.params.blogId);
+  // Find the blog from DB
+  const foundBlog = await Blog.findById(req.params.blogId);
 
   // Blog not found
-  if (!deletedBlog)
+  if (!foundBlog)
     return res.status(404).json({ success: false, message: "Blog not found" });
+
+  // Logged In user should be author of blog
+  if (req.loggedInUser != foundBlog.author)
+    return res.status(401).json({
+      success: false,
+      message: "Unauthorized",
+    });
+
+  // Delete blog from DB
+  await Blog.findByIdAndDelete(req.params.blogId);
 
   return res.status(200).json({ success: true, message: "Blog Deleted" });
 };
